refactor(lfa): move concrete LFA types to hidden internal submodule
